requirements.txt
aiohttp==3.11.18
APScheduler==3.10.4
feedparser==6.0.11
openai==1.78.1
python-dotenv==1.0.1
requests==2.32.3
Pillow==10.4.0
pixivpy3==3.7.5
